Rename ast_parser_python24 to ast_parser_python25
We previously needed this for both Python 2.4 and 2.5, but we no longer
support 2.4, so this makes it clearer that it is still needed.
